requirements.txt
# Core Flask dependencies
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.0
Flask-JWT-Extended==4.4.4
Flask-CORS==3.0.10
flask-restx==1.0.3

# Database
SQLAlchemy==1.4.45
PyMySQL==1.1.0

# Security
Werkzeug==2.2.2
PyJWT==2.6.0

# Configuration
python-decouple==3.6
python-dotenv==0.21.0

# Email dependencies (already included in Python standard library)
# smtplib, ssl, email.mime are built-in modules

# Payment processing
stripe>=5.0.0

# Utilities
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Mako==1.2.4

# Text-to-Speech and Text Processing
edge-tts>=6.1.16
pydantic==1.10.12
beautifulsoup4==4.12.2
markdown==3.5.1
requests==2.31.0
groq>=0.4.0

# Production server
gunicorn==21.2.0

# Dependencies
aniso8601==9.0.1
attrs==22.1.0
greenlet==2.0.1
importlib-metadata==5.1.0
jsonschema==4.17.3
pyrsistent==0.19.2
pytz==2022.6
six==1.16.0
zipp==3.11.0
